moved "generate_manager_config" to utils.py
