+++ api_server/launch_server.py
@@ -7,39 +7,12 @@ import os
 import subprocess
 
 import uvicorn
-import yaml
+
+from skyflow.utils.utils import generate_manager_config
 
 API_SERVER_CONFIG_PATH = "~/.skyconf/config.yaml"
 API_SERVER_HOST = "127.0.0.1"
 API_SERVER_PORT = 50051
-
-
-def generate_manager_config(host: str, port: int):
-    """Generates the API server config file."""
-    absolute_path = os.path.expanduser(API_SERVER_CONFIG_PATH)
-    # If path exists, check if host and port are identical
-    if os.path.exists(absolute_path):
-        with open(absolute_path, "r") as config_file:
-            config_dict = yaml.safe_load(config_file)
-
-            if (config_dict["api_server"]["host"] == host
-                    and config_dict["api_server"]["port"] == port
-                    and "secret" in config_dict["api_server"]):
-                print("API server config already exists. Skipping generation.")
-                return
-
-    config_dict = {
-        "api_server": {
-            "host": host,
-            "port": port,
-            "secret": os.urandom(256).hex(),
-        },
-        "users": [],
-    }
-    os.makedirs(os.path.dirname(absolute_path), exist_ok=True)
-    with open(absolute_path, "w") as config_file:
-        yaml.dump(config_dict, config_file)
-
 
 def check_and_install_etcd(data_directory=None):
     """
