+++ skyflow/utils/utils.py
@@ -14,6 +14,33 @@ API_SERVER_CONFIG_PATH = "~/.skyconf/config.yaml"
 OBJECT_TEMPLATES = importlib.import_module("skyflow.templates")
 
 
+def generate_manager_config(host: str, port: int):
+    """Generates the API server config file."""
+    absolute_path = os.path.expanduser(API_SERVER_CONFIG_PATH)
+    # If path exists, check if host and port are identical
+    if os.path.exists(absolute_path):
+        with open(absolute_path, "r") as config_file:
+            config_dict = yaml.safe_load(config_file)
+
+            if (config_dict["api_server"]["host"] == host
+                    and config_dict["api_server"]["port"] == port
+                    and "secret" in config_dict["api_server"]):
+                print("API server config already exists. Skipping generation.")
+                return
+
+    config_dict = {
+        "api_server": {
+            "host": host,
+            "port": port,
+            "secret": os.urandom(256).hex(),
+        },
+        "users": [],
+    }
+    os.makedirs(os.path.dirname(absolute_path), exist_ok=True)
+    with open(absolute_path, "w") as config_file:
+        yaml.dump(config_dict, config_file)
+
+
 def load_object(response: Union[Dict, List[Dict]]):
     """
     Loads an object or a list of objects (from templates) from a dictionary.
